Extracted actual tests to lionrock-grpc-tests (which now doesn't work)
