+++ lionrock-foundationdb-server/src/test/java/io/github/panghy/lionrock/foundationdb/GrpcTestSuite.java
@@ -1,0 +1,9 @@
+package io.github.panghy.lionrock.foundationdb;
+
+import org.junit.platform.suite.api.SelectPackages;
+import org.junit.platform.suite.api.Suite;
+
+@Suite
+@SelectPackages("io.github.panghy.lionrock.tests")
+public class GrpcTestSuite {
+}
