+++ lionrock-grpc-tests/src/testFixtures/java/io/github/panghy/lionrock/tests/BidirectionalStreamingTests.java
@@ -1,8 +1,7 @@
-package io.github.panghy.lionrock.foundationdb;
+package io.github.panghy.lionrock.tests;
 
 import com.google.protobuf.ByteString;
 import io.github.panghy.lionrock.proto.*;
-import io.github.panghy.lionrock.tests.AbstractStreamingGrpcTest;
 import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
 import io.grpc.stub.StreamObserver;
